requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.7.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.5
Hypercorn==0.17.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
koreanize-matplotlib==0.1.1
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
motor==3.6.1
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==2.0.1
openai==0.28.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
priority==2.0.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.1.2
pytesseract==0.3.13
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
Quart==0.19.6
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.38.2
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
Werkzeug==3.0.3
win32_setctime==1.2.0
wsproto==1.2.0
xxhash==3.4.1
xyzservices==2024.6.0
yarl==1.9.4
